decompose json presenter
